requirements.txt
Jinja2==2.8
requests==2.9.1
iso8601==0.1.10
beautifulsoup4==4.4.1
pillow==3.1.0
feedparser==5.2.1
pyyaml==3.11
pytz==2015.7

# needs libffi-dev
# this is for TLS SNI
ndg-httpsclient==0.4.0

colorlog==2.10.0
